% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipinfo.R
\name{ipinfo}
\alias{ipinfo}
\title{Look Up Information about an IP Address}
\usage{
ipinfo(ip = NULL)
}
\arguments{
\item{ip}{An IP address. If none is provided, your own IP address is used}
}
\value{
A list containing information about the IP address, including the
owner, the hostname (if applicable), and geological information.
}
\description{
\code{ipinfo} looks up information about a given IP address using the
\href{https://ipinfo.io}{ipinfo.io} API.
}
\examples{
\dontrun{
my_ip_info <- ipinfo()
google_dns <- ipinfo(ip = "8.8.8.8")
}
}
